index.tsxを整理しました．
