+++ front-end/front-app/src/pages/components/memoData.ts
@@ -42,5 +42,19 @@ export const memos: MemoData[] = [
     title: "ポケモンダイアモンド・パール",
     comment: "めっちゃ楽しい！映画化したけど見に行こうかな",
 },
+{
+    tag: ["ごはん", "パスタ"],
+    date: 20230817111700,
+    title: "",
+    comment: "久しぶりに家でパスタたべた！",
+},
+{
+    tag: ["ごはん", "パスタ"],
+    date: 20230602111700,
+    title: "",
+    comment: "粉チーズとハバネロ入れるとやっぱそれっぽくなるよね",
+},
+
+
 
 ];
